Large Refactoring
Renaming and decoupled the builder and Node
Refactored Lexer
