+++ src/rowdy/RowdyLexer.java
@@ -77,6 +77,10 @@ public class RowdyLexer {
     fileStack = parseFile(fileName);
   }
 
+  public void parseLine(String code) {
+    
+  }
+  
   /**
    * Gets the contents of a file and adds all relevant groupings of characters
    * to a stack.
@@ -90,96 +94,97 @@ public class RowdyLexer {
     List<String> symbols = new LinkedList<>();
     String line;
     File file = new File(fileName);
-    BufferedReader reader = null;
+    try (BufferedReader reader = new BufferedReader(new FileReader(file))) {
+      while ((line = reader.readLine()) != null) {
+        symbols.addAll(parseCode(line));
+      }
+    }
+    return symbols;
+  }
+  
+  public List<String> parseCode(String line) {
+    List<String> symbols = new LinkedList<>();
     boolean eoln = false;
-
-    reader = new BufferedReader(new FileReader(file));
     char cur;
     String word = "";
-    while ((line = reader.readLine()) != null) {
+    for (int c = 0, prev = 0; c < line.length(); c++) {
+      cur = line.charAt(c);
 
-      for (int c = 0, prev = 0; c < line.length(); c++) {
-        cur = line.charAt(c);
-
-        if (cur == '/') {
-          if (line.charAt(c + 1) == '/') {
+      if (cur == '/') {
+        if (line.charAt(c + 1) == '/') {
+          break;
+        }
+      }
+      if (id.matcher(Character.toString(cur)).matches()) {
+        word += cur;
+        do {
+          c++;
+          if (c >= line.length()) {
+            eoln = true;
             break;
           }
-        }
-        if (id.matcher(Character.toString(cur)).matches()) {
+          cur = line.charAt(c);
           word += cur;
-          do {
-            c++;
-            if (c >= line.length()) {
-              eoln = true;
-              break;
-            }
-            cur = line.charAt(c);
-            word += cur;
-          } while (id.matcher(word).matches());
-        } else if (cur == '\"') {
-          word += cur;
-          do {
-            c++;
-            if (c >= line.length()) {
-              break;
-            }
-            cur = line.charAt(c);
-            word += cur;
-          } while (cur != '\"');
-          eoln = true;
-        } else if (tokenNumber.matcher(Character.toString(cur)).matches()) {
-          word += cur;
-          do {
-            c++;
-            if (c >= line.length()) {
-              eoln = true;// Flag adjustments
-              break;
-            }
-            cur = line.charAt(c);
-            word += cur;
-          } while (number.matcher(word).matches());
-        } else if (SPECIALCHARGROUP.contains(Character.toString(cur))
-                && cur != ' ') {
-          word += cur;
-          do {
-            c++;
-            if (c >= line.length()) {
-              eoln = true;// Flag adjustments
-              break;
-            }
-            cur = line.charAt(c);
-            word += cur;
-          } while (SPECIALCHARGROUP.contains(word) && cur != ' ');
-        }
-
-        if (!word.isEmpty()) {
-          symbols.add(
-                  (eoln)
-                          ? word
-                          :// Else
-                          word.substring(0, word.length() - 1));
-        }
-
-        if (!whiteSpace.matcher(Character.toString(cur)).matches()) {
-          if (c == prev) { // Stop parsing on unknown symbol
-            throw new RuntimeException("Parsing halted, unable"
-                    + " to resolve character '" + cur + "'");
+        } while (id.matcher(word).matches());
+      } else if (cur == '\"') {
+        word += cur;
+        do {
+          c++;
+          if (c >= line.length()) {
+            break;
           }
-          if (c < line.length() && !eoln) {
-            prev = c;
-            c--;
+          cur = line.charAt(c);
+          word += cur;
+        } while (cur != '\"');
+        eoln = true;
+      } else if (tokenNumber.matcher(Character.toString(cur)).matches()) {
+        word += cur;
+        do {
+          c++;
+          if (c >= line.length()) {
+            eoln = true;// Flag adjustments
+            break;
           }
-        }
-        eoln = false;
-        word = "";
+          cur = line.charAt(c);
+          word += cur;
+        } while (number.matcher(word).matches());
+      } else if (SPECIALCHARGROUP.contains(Character.toString(cur))
+              && cur != ' ') {
+        word += cur;
+        do {
+          c++;
+          if (c >= line.length()) {
+            eoln = true;// Flag adjustments
+            break;
+          }
+          cur = line.charAt(c);
+          word += cur;
+        } while (SPECIALCHARGROUP.contains(word) && cur != ' ');
       }
 
-      symbols.add("EOLN");
+      if (!word.isEmpty()) {
+        symbols.add(
+                (eoln)
+                        ? word
+                        :// Else
+                        word.substring(0, word.length() - 1));
+      }
+
+      if (!whiteSpace.matcher(Character.toString(cur)).matches()) {
+        if (c == prev) { // Stop parsing on unknown symbol
+          throw new RuntimeException("Parsing halted, unable"
+                  + " to resolve character '" + cur + "'");
+        }
+        if (c < line.length() && !eoln) {
+          prev = c;
+          c--;
+        }
+      }
+      eoln = false;
+      word = "";
     }
-    if (reader != null) {
-      reader.close();
-    }
+
+    symbols.add("EOLN");
     return symbols;
   }
