--- test/rowdy/RowdyParseTreeTest.java
@@ -1,35 +1,0 @@
-
-package rowdy;
-
-import org.junit.*;
-
-
-/**
- *
- * @author Richard
- */
-public class RowdyParseTreeTest {
-  
-  /* TODO
-  
-    1) Mock Tokenizer and Language
-  
-  */
-  
-  private RowdyParseTree parseTree;
-  
-  @Before
-  public void setUp(){
-    
-  }
-  
-  @After
-  public void tearDown(){
-    
-  }
-  @Test
-  public void test(){
-    
-  }
-
-}
